+++ core/src/main/java/tech/ydb/core/impl/priority/LocalDCPriorityEndpointEvaluator.java
@@ -11,18 +11,23 @@ public class LocalDCPriorityEndpointEvaluator implements PriorityEndpointEvaluat
 
     private final String localDCLocation;
 
+    private volatile String preferred;
+
     public LocalDCPriorityEndpointEvaluator(String localDCLocation) {
         this.localDCLocation = localDCLocation;
     }
 
     @Override
-    public long evaluatePriority(String selfLocation, DiscoveryProtos.EndpointInfo endpointInfo) {
-        String preferred = localDCLocation;
+    public long evaluatePriority(DiscoveryProtos.EndpointInfo endpointInfo) {
+        return preferred.equalsIgnoreCase(endpointInfo.getLocation()) ? 0 : LOCALITY_SHIFT;
+    }
+
+    @Override
+    public void prepareStatement(DiscoveryProtos.ListEndpointsResult result) {
+        preferred = localDCLocation;
 
         if (preferred == null || preferred.isEmpty()) {
-            preferred = selfLocation;
+            preferred = result.getSelfLocation();
         }
-
-        return preferred.equalsIgnoreCase(endpointInfo.getLocation()) ? 0 : LOCALITY_SHIFT;
     }
 }