+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.1.0 ##
 
+* Added support detect local dc
 * Support creating asynchronous index in createTable and describeTable
 * Fixed default timeout for readTable
 * Fixed exponential backoff calculation